# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatFlow API - Sistema de Atendimento
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  agent: "agent",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Contact {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  company?: string | null;
  location?: string | null;
  avatarUrl?: string | null;
  conversationsCount: number;
  createdAt: string;
}

export interface CreateContactRequest {
  name: string;
  email?: string | null;
  phone?: string | null;
  company?: string | null;
  location?: string | null;
}

export interface ContactsListResponse {
  contacts: Contact[];
  total: number;
  page: number;
  limit: number;
}

export type ConversationStatus =
  (typeof ConversationStatus)[keyof typeof ConversationStatus];

export const ConversationStatus = {
  open: "open",
  resolved: "resolved",
  pending: "pending",
  snoozed: "snoozed",
} as const;

export type ConversationChannel =
  (typeof ConversationChannel)[keyof typeof ConversationChannel];

export const ConversationChannel = {
  email: "email",
  chat: "chat",
  whatsapp: "whatsapp",
  telegram: "telegram",
  api: "api",
} as const;

export type ConversationPriority =
  | (typeof ConversationPriority)[keyof typeof ConversationPriority]
  | null;

export const ConversationPriority = {
  none: "none",
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export type MessageMessageType =
  (typeof MessageMessageType)[keyof typeof MessageMessageType];

export const MessageMessageType = {
  incoming: "incoming",
  outgoing: "outgoing",
  activity: "activity",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  content: string;
  messageType: MessageMessageType;
  sender?: User | null;
  createdAt: string;
}

export interface Conversation {
  id: number;
  status: ConversationStatus;
  channel: ConversationChannel;
  priority?: ConversationPriority;
  subject?: string | null;
  unreadCount: number;
  contact: Contact;
  assignee?: User | null;
  lastMessage?: Message | null;
  labels: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateConversationRequestChannel =
  (typeof CreateConversationRequestChannel)[keyof typeof CreateConversationRequestChannel];

export const CreateConversationRequestChannel = {
  email: "email",
  chat: "chat",
  whatsapp: "whatsapp",
  telegram: "telegram",
  api: "api",
} as const;

export interface CreateConversationRequest {
  contactId: number;
  channel: CreateConversationRequestChannel;
  subject?: string | null;
  assigneeId?: number | null;
}

export type UpdateConversationRequestStatus =
  (typeof UpdateConversationRequestStatus)[keyof typeof UpdateConversationRequestStatus];

export const UpdateConversationRequestStatus = {
  open: "open",
  resolved: "resolved",
  pending: "pending",
  snoozed: "snoozed",
} as const;

export type UpdateConversationRequestPriority =
  | (typeof UpdateConversationRequestPriority)[keyof typeof UpdateConversationRequestPriority]
  | null;

export const UpdateConversationRequestPriority = {
  none: "none",
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateConversationRequest {
  status?: UpdateConversationRequestStatus;
  assigneeId?: number | null;
  priority?: UpdateConversationRequestPriority;
  labels?: string[];
}

export type SendMessageRequestMessageType =
  (typeof SendMessageRequestMessageType)[keyof typeof SendMessageRequestMessageType];

export const SendMessageRequestMessageType = {
  incoming: "incoming",
  outgoing: "outgoing",
} as const;

export interface SendMessageRequest {
  content: string;
  messageType: SendMessageRequestMessageType;
}

export type ListConversationsParams = {
  status?: ListConversationsStatus;
  assigneeId?: number;
  search?: string;
};

export type ListConversationsStatus =
  (typeof ListConversationsStatus)[keyof typeof ListConversationsStatus];

export const ListConversationsStatus = {
  open: "open",
  resolved: "resolved",
  pending: "pending",
  snoozed: "snoozed",
} as const;

export type ListContactsParams = {
  search?: string;
  page?: number;
  limit?: number;
};