# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChatFlow API - Sistema de Atendimento
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: conversations
    description: Conversation operations
  - name: contacts
    description: Contact operations
  - name: messages
    description: Message operations
  - name: users
    description: User operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List all conversations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, pending, snoozed]
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a single conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
    put:
      operationId: updateConversation
      tags: [conversations]
      summary: Update a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConversationRequest"
      responses:
        "200":
          description: Conversation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
    delete:
      operationId: deleteConversation
      tags: [conversations]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Contacts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a single contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  /contacts/{id}/conversations:
    get:
      operationId: listContactConversations
      tags: [contacts]
      summary: List conversations for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users/agents
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - name
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent]
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt

    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        conversationsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - conversationsCount
        - createdAt

    CreateContactRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required:
        - name

    ContactsListResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - contacts
        - total
        - page
        - limit

    Conversation:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [open, resolved, pending, snoozed]
        channel:
          type: string
          enum: [email, chat, whatsapp, telegram, api]
        priority:
          type: string
          enum: [none, low, medium, high, urgent]
          nullable: true
        subject:
          type: string
          nullable: true
        unreadCount:
          type: integer
        contact:
          $ref: "#/components/schemas/Contact"
        assignee:
          $ref: "#/components/schemas/User"
          nullable: true
        lastMessage:
          $ref: "#/components/schemas/Message"
          nullable: true
        labels:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - channel
        - unreadCount
        - contact
        - labels
        - createdAt
        - updatedAt

    CreateConversationRequest:
      type: object
      properties:
        contactId:
          type: integer
        channel:
          type: string
          enum: [email, chat, whatsapp, telegram, api]
        subject:
          type: string
          nullable: true
        assigneeId:
          type: integer
          nullable: true
      required:
        - contactId
        - channel

    UpdateConversationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, resolved, pending, snoozed]
        assigneeId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [none, low, medium, high, urgent]
          nullable: true
        labels:
          type: array
          items:
            type: string

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        content:
          type: string
        messageType:
          type: string
          enum: [incoming, outgoing, activity]
        sender:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - content
        - messageType
        - createdAt

    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
        messageType:
          type: string
          enum: [incoming, outgoing]
      required:
        - content
        - messageType