# lib/api-zod/src/generated/types/createContactRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatFlow API - Sistema de Atendimento
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactRequest {
  name: string;
  email?: string | null;
  phone?: string | null;
  company?: string | null;
  location?: string | null;
}