# lib/api-zod/src/generated/types/updateConversationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatFlow API - Sistema de Atendimento
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateConversationRequestPriority } from "./updateConversationRequestPriority";
import type { UpdateConversationRequestStatus } from "./updateConversationRequestStatus";

export interface UpdateConversationRequest {
  status?: UpdateConversationRequestStatus;
  assigneeId?: number | null;
  priority?: UpdateConversationRequestPriority;
  labels?: string[];
}